Внес правки по ревью
