--- app/src/main/java/com/kponomarev/oneofthenewproject/Crypto.kt
@@ -1,5 +1,0 @@
-package com.kponomarev.oneofthenewproject
-
-data class Crypto(
-    val name: String
-)
