--- app/src/main/java/com/kponomarev/oneofthenewproject/Fragment3.kt
@@ -1,22 +1,0 @@
-package com.kponomarev.oneofthenewproject
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class Fragment3 : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_3, container, false)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() = Fragment3()
-            }
-}
