--- app/src/main/java/com/kponomarev/oneofthenewproject/MainActivity.kt
@@ -1,34 +1,0 @@
-package com.kponomarev.oneofthenewproject
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import by.kirich1409.viewbindingdelegate.viewBinding
-import com.google.android.material.tabs.TabLayout
-import com.kponomarev.oneofthenewproject.databinding.ActivityMainBinding
-
-@Suppress("UnusedPrivateMember")
-class MainActivity : AppCompatActivity(R.layout.activity_main) {
-    private val binding: ActivityMainBinding by viewBinding()
-    private val fragList = listOf(
-        CryptoFragment.newInstance(),
-        Fragment2.newInstance(),
-        Fragment3.newInstance()
-    )
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-        binding.tabLayout.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
-            override fun onTabSelected(tab: TabLayout.Tab?) {
-                supportFragmentManager.beginTransaction()
-                    .replace(R.id.placeHolder, fragList[tab?.position!!]).commit()
-            }
-
-            override fun onTabUnselected(tab: TabLayout.Tab?) {
-            }
-
-            override fun onTabReselected(tab: TabLayout.Tab?) {
-            }
-        })
-    }
-}
