--- app/src/main/res/layout/fragment_crypto.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    tools:context=".CryptoFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/cryptoRcView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:listitem="@layout/crypto_list_item" />
-
-</FrameLayout>
